requirements.txt
aniso8601==7.0.0
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.2
Django==3.2.5
django-cors-headers==3.7.0
django-filter==2.4.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
drf-yasg==1.20.0
Faker==8.11.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mixer==7.1.2
packaging==21.0
pipenv==2021.5.29
platformdirs==2.2.0
promise==2.3
psycopg2-binary==2.9.1
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
ruamel.yaml==0.17.11
ruamel.yaml.clib==0.2.6
Rx==1.6.1
singledispatch==3.6.2
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.26.6
virtualenv==20.7.2
virtualenv-clone==0.5.6